refactor(transaction): move transactions to data folder
