+++ FinEdgeFrontend/src/Data/Transactions/Transactions.tsx
@@ -1,3 +1,4 @@
+import CategoriesSankeyChart from '../../Components/CategoriesPage/CategoriesSankeyChart';
 import CreateTransaction from './CreateTransaction';
 import GetTransactions from './GetTransactions';
 import TransactionReports from './TransactionReports';
@@ -6,6 +7,7 @@ const Transactions = () => {
   return (
     <>
       <CreateTransaction />
+      <CategoriesSankeyChart />
       <GetTransactions />
       <TransactionReports />
     </>
